carga-datos/requirements.txt
requests
pandas
fastapi
python-multipart
uvicorn
psycopg2-binary